Optionally generate ID attributes for heading nodes
In order to do this, all render functions are expected to now return a
Phoenix.HTML.Safe, or they're treated as escaped text. This is a breaking
change, but the right thing to do for safety.

Added a render_document function that converts to a String in the final step,
for convenience.